// my_frontend/src/types.tsx
export interface Employee {
     id: number;
     full_name: string;
     email: string;
     role: string;
          
}

export interface Project {
     id: number;
     project_name: string; 
     date_created: string;
     description: string;
          
}

export interface Department {
     id: number;
     department_name: string;
     location: string;
  
}